feat: adding acs autoresolve
