+++ megamek/i18n/megamek/common/acs-report-messages.properties
@@ -1,0 +1,18 @@
+
+1000=ACS Auto Resolve Scenario
+1001=Round <data>
+1002=Scenario Resolved at the end of round <data>
+
+1100=Initiative Phase
+1101=<data> rolls <data> for initiative.
+1102=Initiative order result is: <data>
+
+1200=Deployment Phase
+
+1300=Detection and Reconnaissance Phase
+
+1400=Movement Phase
+
+1500=Combat Phase
+
+1600=End Phase
