+++ megamek/i18n/megamek/common/acs-report-messages_en.properties
@@ -1,0 +1,56 @@
+#
+# Copyright (c) 2024 - The MegaMek Team. All Rights Reserved.
+#
+# This file is part of MegaMek.
+#
+# MegaMek is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# MegaMek is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
+#
+
+1000=Initiative Phase for Round {0}
+1005=Initiative Phase for Start of Game Deployment
+1010=Initiative Phase for Deployment for Round {0}
+1015=\ rolls a
+1020=The general turn order is:
+1025=Wind direction is from the {0}.
+1030=Wind strength is {0}.
+1031=The weather is {0}.
+1032=Visibility is {0}.
+1033=Fog level is {0}.
+1035=<newline><B>Targeting Phase</B><newline>-------------------
+1040=<newline>The turn order is:
+1045=<data> (<data>), rolled <data>
+1050=<data> (special turn)
+1060=<newline>Remaining units to deploy:
+1065=Round: #<data>
+1066=<data> (<data>), Start: <data>
+1100=<newline><B>Offboard Attack Phase</B><newline>-------------------
+1200=-------------------
+1205=<nothing>
+1210=
+1215=<newline><newline><newline>A nuclear weapon goes off in hex <data>!!!
+1216=<span class='info'>End of nuclear resolution</span>
+1220=<span class='info'>End of nuclear blast effects</span><newline>
+1230=<data>
+1231=<data> (<data>) <data>
+1240=<data> (<data>) needs to roll <data> or greater to avoid Blue Shield Failure, rolls <data>: <msg:1241,1242>
+1241=Blue Shield Field Dampener <span class='warning'>fails</span>!
+1242=no effect
+
+2001={0} attacks {1}.
+2003=The attack requires a {0}.
+
+2010=Attack cannot succeed ({0}).
+2012=Attack missed.
+2013=The attack hits!
+2020={0} rolls a {1}.
