+++ megamek/src/megamek/common/IGame.java
@@ -26,7 +26,10 @@ import megamek.common.force.Forces;
 import megamek.common.options.BasicGameOptions;
 import megamek.server.scriptedevent.TriggeredEvent;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -328,7 +331,6 @@ public interface IGame {
     /**
      * This is a Client-side method to replace or add units that are sent from the
      * server.
-     *
      * Adds the given units to the list of units or objects in the current game.
      * When a unit's ID is already
      * present the currently assigned unit will be replaced with the given new one.
@@ -343,6 +345,13 @@ public interface IGame {
      *         should have a reason for their removal set.
      */
     List<InGameObject> getGraveyard();
+
+    /**
+     * @return the number of non-destroyed entities owned by the player, including
+     *         entities not yet
+     *         deployed. Ignores offboard units and captured Mek pilots.
+     */
+    int getLiveDeployedEntitiesOwnedBy(Player player);
 
     // endregion
 
@@ -446,5 +455,51 @@ public interface IGame {
      */
     void addScriptedEvent(TriggeredEvent event);
 
-    // endregion
+
+    /**
+     * Setter for property victoryPlayerId.
+     *
+     * @param victoryPlayerId New value of property victoryPlayerId.
+     */
+    void setVictoryPlayerId(int victoryPlayerId);
+
+    /**
+     * Setter for property victoryPlayerId.
+     *
+     * @param victoryTeam New value of property victoryPlayerId.
+     */
+    void setVictoryTeam(int victoryTeam);
+
+    /**
+     * Cancels a victory
+     */
+    void cancelVictory();
+
+
+    /**
+     * Getter for property victoryPlayerId.
+     *
+     * @return Value of property victoryPlayerId.
+     */
+    int getVictoryPlayerId();
+
+    /**
+     * Getter for property victoryTeam.
+     *
+     * @return Value of property victoryTeam.
+     */
+    int getVictoryTeam();
+
+    /**
+     * Verify if the game has ended due a time limit
+     * @return true if the game has ended due to a time limit
+     */
+    boolean gameTimerIsExpired();
+
+    /**
+     * Get how many commanders still alive are owned by the player
+     * @param player the object player
+     * @return the number of commanders owned by the player
+     */
+    int getLiveCommandersOwnedBy(Player player);
 }
