+++ megamek/src/megamek/server/victory/BVDestroyedVictoryCondition.java
@@ -20,6 +20,7 @@
 package megamek.server.victory;
 
 import megamek.common.Game;
+import megamek.common.IGame;
 import megamek.common.Player;
 import megamek.common.Report;
 
@@ -39,7 +40,7 @@ public class BVDestroyedVictoryCondition implements BvVictoryCondition {
     }
 
     @Override
-    public VictoryResult checkVictory(Game game, Map<String, Object> ctx) {
+    public VictoryResult checkVictory(IGame game, Map<String, Object> ctx) {
         boolean isVictory = false;
         VictoryResult victoryResult = new VictoryResult(true);
         HashSet<Integer> doneTeams = new HashSet<>();
