+++ megamek/src/megamek/server/victory/BvVictoryCondition.java
@@ -20,6 +20,7 @@
 package megamek.server.victory;
 
 import megamek.common.Game;
+import megamek.common.IGame;
 import megamek.common.Player;
 
 import java.io.Serializable;
@@ -29,21 +30,21 @@ import java.io.Serializable;
  */
 public interface BvVictoryCondition extends VictoryCondition, Serializable {
 
-    default int getFriendlyBV(Game game, Player player) {
+    default int getFriendlyBV(IGame game, Player player) {
         return game.getPlayersList().stream()
                 .filter(Player::isNotObserver)
                 .filter(p -> !p.isEnemyOf(player))
                 .mapToInt(Player::getBV).sum();
     }
 
-    default int getEnemyBV(Game game, Player player) {
+    default int getEnemyBV(IGame game, Player player) {
         return game.getPlayersList().stream()
                 .filter(Player::isNotObserver)
                 .filter(p -> p.isEnemyOf(player))
                 .mapToInt(Player::getBV).sum();
     }
 
-    default int getEnemyInitialBV(Game game, Player player) {
+    default int getEnemyInitialBV(IGame game, Player player) {
         return game.getPlayersList().stream()
                 .filter(Player::isNotObserver)
                 .filter(p -> p.isEnemyOf(player))
