+++ megamek/src/megamek/server/victory/PlayerAgreedVictory.java
@@ -20,6 +20,7 @@
 package megamek.server.victory;
 
 import megamek.common.Game;
+import megamek.common.IGame;
 import megamek.common.Player;
 
 import java.io.Serializable;
@@ -32,10 +33,12 @@ import java.util.Map;
 public class PlayerAgreedVictory implements VictoryCondition, Serializable {
 
     @Override
-    public VictoryResult checkVictory(Game game, Map<String, Object> ctx) {
+    public VictoryResult checkVictory(IGame game, Map<String, Object> ctx) {
         if (!game.isForceVictory()) {
             return VictoryResult.noResult();
         }
+        // Set this way because currently the only type of "Game" that you can have players
+        // agree to a victory is when running the mm client, which implements Game
         final int victoryPlayerId = game.getVictoryPlayerId();
         final int victoryTeam = game.getVictoryTeam();
         List<Player> activePlayers = game.getPlayersList();
