+++ megamek/src/megamek/server/victory/VictoryResult.java
@@ -20,6 +20,7 @@
 package megamek.server.victory;
 
 import megamek.common.Game;
+import megamek.common.IGame;
 import megamek.common.Player;
 import megamek.common.Report;
 
@@ -43,7 +44,7 @@ public final class VictoryResult {
     public VictoryResult(boolean win) {
         this.isVictory = win;
     }
-    
+
     public VictoryResult(boolean win, int player, int team) {
         this.isVictory = win;
         if (player != Player.PLAYER_NONE) {
@@ -53,11 +54,11 @@ public final class VictoryResult {
             setTeamScore(team, 1.0);
         }
     }
-    
+
     public static VictoryResult noResult() {
         return new VictoryResult(false, Player.PLAYER_NONE, Player.TEAM_NONE);
     }
-    
+
     public static VictoryResult drawResult() {
         return new VictoryResult(true, Player.PLAYER_NONE, Player.TEAM_NONE);
     }
@@ -157,7 +158,7 @@ public final class VictoryResult {
      * @param game The current {@link Game} for which the victory needs to be processed
      * @return a list of reports generated by victory checking and by the actual processing of the victory
      */
-    public List<Report> processVictory(Game game) {
+    public List<Report> processVictory(IGame game) {
         List<Report> gatheredReports = getReports();
         if (isVictory()) {
             boolean draw = isDraw();
